- Suppress compilation warnings with MSVC++9.0
git-svn-id: http://root.cern.ch/svn/root/trunk@25744 27541ba8-7e3a-0410-8455-c3a389f83636

+++ graf2d/win32gdk/src/gifquantize.c
@@ -147,9 +147,9 @@ int GIFquantize(unsigned int Width, unsigned int Height, int *ColorMapSize,
                 Blue += QuantizedColor -> RGB[2];
                 QuantizedColor = QuantizedColor -> Pnext;
             }
-            OutputColorMap[i].Red = (Red << (8 - BITS_PER_PRIM_COLOR)) / j;
-            OutputColorMap[i].Green = (Green << (8 - BITS_PER_PRIM_COLOR)) / j;
-            OutputColorMap[i].Blue= (Blue << (8 - BITS_PER_PRIM_COLOR)) / j;
+            OutputColorMap[i].Red = (byte)((Red << (8 - BITS_PER_PRIM_COLOR)) / j);
+            OutputColorMap[i].Green = (byte)((Green << (8 - BITS_PER_PRIM_COLOR)) / j);
+            OutputColorMap[i].Blue= (byte)((Blue << (8 - BITS_PER_PRIM_COLOR)) / j);
         }
         else
             fprintf(stderr, "Null entry in quantized color map - thats weird.");
